backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
crawl4ai==0.1.0
supabase==0.0.2
mistralai==0.0.7
python-dotenv==0.19.0
numpy==1.21.2
uuid
PyJWT==2.8.0